requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
chroma-hnswlib==0.7.3
chromadb==0.5.0
langchain==0.2.5
langchain_cohere==0.1.8
langchain_community==0.2.5
langchain_core==0.2.9
langchain_openai==0.1.9
langchain_text_splitters==0.2.1
langgraph==0.1.1
Requests==2.32.3
tiktoken==0.7.0
typing_extensions==4.12.2
python-dotenv==1.0.1

